requirements.txt
black==22.1.0
coverage==6.3.1
faker==12.3.0
flask-sqlalchemy==2.5.1
flask-swagger-ui==3.36.0
jsonschema==4.4.0
pip-chill==1.0.1
psycopg2-binary==2.9.3
pytest==7.0.0
python-dotenv==0.19.2
